backend/requirements.txt
Flask
Flask-Cors
Unidecode
strawberry-graphql